fix Require cycle
