+++ app/utils/theme/index.ts
@@ -7,8 +7,8 @@ import tinyColor from 'tinycolor2';
 
 import {Preferences} from '@constants';
 import {MODAL_SCREENS_WITHOUT_BACK} from '@constants/screens';
-import {appearanceControlledScreens, mergeNavigationOptions} from '@screens/navigation';
 import EphemeralStore from '@store/ephemeral_store';
+import {appearanceControlledScreens, mergeNavigationOptions} from '@utils/navigation';
 
 import type {Options} from 'react-native-navigation';
